backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.12.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.73.0
greenlet==3.0.3
groq==0.9.0
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.27.0
idna==3.11
ImageHash==4.3.1
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.3
opencv-python==4.10.0.84
orjson==3.11.7
pillow==11.1.0
playwright==1.44.0
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.7.4
pydantic_core==2.18.4
pyee==11.1.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-whois==0.9.5
PyWavelets==1.9.0
PyYAML==6.0.3
rapidfuzz==3.9.3
requests==2.32.3
rich==14.3.3
rich-toolkit==0.19.7
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.37.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.1
watchfiles==1.1.1
websockets==16.0

# fastapi==0.111.0
# uvicorn==0.30.1
# python-dotenv==1.0.1
# requests==2.32.3
# python-whois==0.9.5
# rapidfuzz==3.9.3
# beautifulsoup4==4.12.3
# playwright==1.44.0
# opencv-python==4.10.0.84
# imagehash==4.3.1
# Pillow==11.1.0
# google-generativeai==0.7.2
# groq==0.9.0
# pydantic==2.7.4
# httpx==0.27.0
# python-multipart==0.0.9